Finish fixing categories and rename method in HtlView.php to follow J!
