+++ src/administrator/components/com_kunena/View/attachments/HtmlView.php
@@ -63,17 +63,19 @@ class HtmlView extends BaseHtmlView
 		$this->listOrdering     = $this->escape($this->state->get('list.ordering'));
 		$this->listDirection    = $this->escape($this->state->get('list.direction'));
 
-		$this->setToolbar();
+		$this->addToolbar();
 
 		return parent::display($tpl);
 	}
 
 	/**
+	 * Add the page title and toolbar.
+	 * 
 	 * @return  void
 	 *
 	 * @since   Kunena 6.0
 	 */
-	protected function setToolbar()
+	protected function addToolbar()
 	{
 		$help_url = 'https://docs.kunena.org/en/manual/backend/attachments';
 		ToolbarHelper::help('COM_KUNENA', false, $help_url);
