+++ src/administrator/components/com_kunena/View/badwords/HtmlView.php
@@ -48,7 +48,8 @@ class HtmlView extends BaseHtmlView
 		$document = Factory::getApplication()->getDocument();
 		$document->setTitle(Text::_('COM_KUNENA_A_BADWORDS_MANAGER'));
 
-		$this->setToolbar();
+		$this->addToolbar();
+
 		return parent::display($tpl);
 	}
 
@@ -59,7 +60,7 @@ class HtmlView extends BaseHtmlView
 	 *
 	 * @since   Kunena 6.0
 	 */
-	protected function setToolbar()
+	protected function addToolbar()
 	{
 		// Get the toolbar object instance
 		$bar = Toolbar::getInstance('toolbar');
