+++ src/administrator/components/com_kunena/View/labels/HtmlView.php
@@ -47,7 +47,7 @@ class HtmlView extends BaseHtmlView
 		$document = Factory::getApplication()->getDocument();
 		$document->setTitle(Text::_('Forum Labels'));
 
-		$this->setToolbar();
+		$this->addToolbar();
 		return parent::display($tpl);
 	}
 
@@ -58,7 +58,7 @@ class HtmlView extends BaseHtmlView
 	 *
 	 * @since   Kunena 6.0
 	 */
-	protected function setToolbar()
+	protected function addToolbar()
 	{
 		// Get the toolbar object instance
 		$bar = Toolbar::getInstance('toolbar');
