+++ src/administrator/components/com_kunena/View/tools/HtmlView.php
@@ -41,6 +41,8 @@ class HtmlView extends BaseHtmlView
 	}
 
 	/**
+	 * Add the page title and toolbar.
+	 * 
 	 * @return  void
 	 *
 	 * @since   Kunena 6.0
