+++ src/administrator/components/com_kunena/View/user/HtmlView.php
@@ -168,17 +168,19 @@ class HtmlView extends BaseHtmlView
 
 		$this->config = \KunenaConfig::getInstance();
 
-		$this->setToolBarEdit();
+		$this->addToolbar();
 
 		return parent::display($tpl);
 	}
 
 	/**
+	 * Add the page title and toolbar.
+	 * 
 	 * @return  void
 	 *
 	 * @since   Kunena 6.0
 	 */
-	protected function setToolBarEdit()
+	protected function addToolbar()
 	{
 		// Set the titlebar text
 		ToolbarHelper::title(Text::_('COM_KUNENA'), 'users');
